Arel::Nodes::Lateral
