+++ lib/arel/extensions.rb
@@ -16,6 +16,7 @@ require 'arel/extensions/indirection'
 require 'arel/extensions/bit_string'
 require 'arel/extensions/natural_join'
 require 'arel/extensions/cross_join'
+require 'arel/extensions/lateral'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
